+++ packages/arb-bridge-eth/contracts/challenge/IChallengeFactory.sol
@@ -24,7 +24,7 @@ interface IChallengeFactory {
         address payable _challenger,
         uint256 _challengePeriodTicks,
         bytes32 _challengeHash,
-        uint challengeType
+        uint256 challengeType
     )
         external
         returns(address);
