+++ packages/arb-bridge-eth/contracts/libraries/SigUtils.sol
@@ -21,13 +21,13 @@ library SigUtils {
 
     function parseSignature(
         bytes memory _signatures,
-        uint _pos
+        uint256 _pos
     )
         public
         pure
         returns (uint8 v, bytes32 r, bytes32 s)
     {
-        uint offset = _pos * 65;
+        uint256 offset = _pos * 65;
         // The signature format is a compact form of:
         //   {bytes32 r}{bytes32 s}{uint8 v}
         // Compact means, uint8 is not padded to 32 bytes.
@@ -70,11 +70,11 @@ library SigUtils {
         uint8 v;
         bytes32 r;
         bytes32 s;
-        uint count = countSignatures(_signatures);
+        uint256 count = countSignatures(_signatures);
         address[] memory addresses = new address[](count);
         bytes memory prefix = "\x19Ethereum Signed Message:\n32";
         bytes32 prefixedHash = keccak256(abi.encodePacked(prefix, _messageHash));
-        for (uint i = 0; i < count; i++) {
+        for (uint256 i = 0; i < count; i++) {
             (v, r, s) = parseSignature(_signatures, i);
             addresses[i] = ecrecover(
                 prefixedHash,
